feat: refactor contract context
