+++ pages/_app.tsx
@@ -6,7 +6,6 @@ import SignTxsModal from "components/SignTxsModal";
 import TxsToastList from "components/TxsToastList";
 import { DAPP_CONFIG } from "const/dappConfig";
 import { ENVIRONMENT } from "const/env";
-import { ContractsProvider } from "context/contracts";
 import { useRecoilAdapter } from "hooks/useRecoilAdapter";
 import { useRefreshAfterTxCompleted } from "hooks/useRefreshAfterTxCompleted";
 import useSentryUser from "hooks/useSentryUser";
@@ -92,14 +91,12 @@ function MyApp({ Component, pageProps }: AppPropsWithLayout) {
                     // completedTransactionsDelay={500}
                 >
                     <>
-                        <ContractsProvider>
-                            <GlobalHooks />
-                            <ProductionErrorBoundary>
-                                {/* <Component {...pageProps} /> */}
-                                {getLayout(<Component {...pageProps} />)}
-                                <ConnectWalletModal />
-                            </ProductionErrorBoundary>
-                        </ContractsProvider>
+                        <GlobalHooks />
+                        <ProductionErrorBoundary>
+                            {/* <Component {...pageProps} /> */}
+                            {getLayout(<Component {...pageProps} />)}
+                            <ConnectWalletModal />
+                        </ProductionErrorBoundary>
                         <div className="fixed bottom-24 left-6 right-6 sm:bottom-12 sm:left-auto sm:right-12 z-toast flex flex-col items-end sm:max-w-[480px] space-y-2 sm:space-y-4">
                             <SignTxNotification />
                             <SignTxsModal />
